requirements.txt
adjustText==1.3.0
altair==6.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cymem==2.0.13
dj-database-url==3.1.2
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docx2txt==0.9
et_xmlfile==2.0.0
fastapi==0.121.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
greenlet==3.2.4
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langdetect==1.0.9
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
murmurhash==1.0.15
narwhals==2.13.0
nltk==3.9.2
numpy==2.3.5
ollama==0.6.1
openpyxl==3.1.5
openai==1.42.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
preshed==3.0.12
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.7
pyparsing==3.2.5
pypdf==6.4.1
PyPDF2==3.0.1
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.6
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlparse==0.5.3
srsly==2.5.2
starlette==0.49.3
streamlit==1.52.1
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.1.1
